==> 02-easy-bounds.r <==
#%% ABOUTME: Visualizes HLZ extrapolation intuition figure used in the paper. ===
# Inputs:
#   - config-and-functions.r (project utilities and theme definitions)
#   - data/emp_data.Rdata (generated by 01-prep-data.r; provides CLZ summaries)
# Outputs:
#   - results/hlz-intuition.pdf
# How to run:
#   Rscript 02-easy-bounds.r

#%% Setup -------------------------------------------------------------------

rm(list = ls())

library(here)
here::i_am("02-easy-bounds.r")

source(here("config-and-functions.r"))

paths <- project_paths()
data_dir <- paths$data
results_dir <- paths$results

load(file.path(data_dir, "emp_data.Rdata"))

#%% Extrapolated HLZ ------------------------------------------------------

## define distributions ====
# empirical dist
F_cz = ecdf(cz_sum$tabs)
# hlz extrap
F_fit = function(tabs) pgamma(tabs, shape = 1, scale = 2)
# data mined
F_dm = ecdf(clz_sum$tabs)

## make plotting data
# define scaling
n_cz = length(cz_sum$tabs)
edge_cz = seq(0,10,0.5)
edge_fit = seq(0,10,0.1)
edge_dm = seq(0,10,0.5)

# scaled data
plotme1 = make_dist_dat(
  F_cz, edge_cz, n_cz, F_fit, edge_fit, n_cz, x_match = c(3.0,Inf), showplot = T
) 
plotme2 = make_dist_dat(
  F_cz, edge_cz, n_cz, F_dm, edge_dm, n_cz, x_match = c(3.0,Inf), showplot = F 
) %>% 
  filter(group==2) %>% 
  mutate(group=3)
plotme = plotme1 %>% rbind(plotme2) %>% 
  mutate(group = factor(group, levels = c(1,2,3)
  ))

# numbers for annotations
h_disc = 2
Pr_disc = 1-F_fit(h_disc)
fdrmax = 0.05/Pr_disc

groupdat = tibble(
  group = c('Exponential w/ Mean = 2','Data-Mined (CLZ)')
  , color = c(MATBLUE, MATRED)
  , linetype = c('solid', 'dashed')
)

## actual plot ====
plt = ggplot(data = plotme, aes(x=mids, y=dF)) +
  geom_bar(
    data = plotme %>% filter(group == 1)
    , stat = 'identity', position = 'identity'
    , aes(fill = group)
  ) +
  geom_line(
    data = plotme %>% filter(group %in% c(2,3))
    , aes(color = group, linetype = group)
  ) +
  scale_fill_manual(values = 'gray', labels = 'Published                ', name = NULL) +  
  scale_color_manual(values = groupdat$color, labels = groupdat$group, name = NULL) +
  scale_linetype_manual(values = groupdat$linetype, labels = groupdat$group, name = NULL)  +
  geom_vline(xintercept = h_disc) +
  xlab(TeX('Absolute t-statistic ($|t_i|$)')) +
  ylab('Number of Signals') +
  scale_x_continuous(
    breaks = seq(0,14,2)
  ) +
  # discovery line
  geom_vline(xintercept = h_disc, color = 'black') +
  theme(
    legend.position = c(70,60)/100
    , legend.margin = margin(t = -15, r = 20, b = 0, l = 5),
  ) +
  coord_cartesian(
    xlim = c(0,8), ylim = c(0, 150)
  ) 

ggsave(file.path(results_dir, "hlz-intuition.pdf"), scale = 1, height = 2.5, width = 5, device = cairo_pdf)
